\name{facorization}
\alias{factorize}

\title{Factorize a number}

\description{Give all primes numbers to factor the number}

\usage{factorize(n)}

\arguments{
  \item{n}{Integer}
}

\value{
  Vector of all numbers.
}

\references{Gnu MP Library see http://swox.com/gmp}
\author{Antoine Lucas}



\examples{
factorize(34455342)
}

\keyword{arith}

