% Generated by roxygen2: do not edit by hand
\name{gmp_plot}
\alias{gmp_plot}
\alias{plot.gmp_plot}
\alias{print.gmp_plot}
\title{Methods for GMP plot objects}
\usage{
\method{plot}{gmp_plot}(x, newpage = TRUE, ...)

\method{print}{gmp_plot}(x, ...)
}
\arguments{
\item{x}{A \code{gmp_plot} object.}
\item{newpage}{Logical; start a new grid graphics page before drawing.}
\item{...}{Additional arguments reserved for future extensions.}
}
\value{Both methods return \code{x} invisibly. \code{plot()} also renders the chromosome map.}
\description{Print or render a reusable \code{gmp_plot} object created by \code{GMP()} or \code{plot_genome_map()}.}
