% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMP.R
\name{GMP}
\alias{GMP}
\title{title}
\usage{
GMP(chr_length = NULL, gene_info = NULL, radian = 0)
}
\arguments{
\item{chr_length}{Chromosome length information}

\item{gene_info}{Gene information}

\item{radian}{Chromosome drawing radian}
}
\description{
title
}
\examples{
library(GMP)
chr_length_default <- data.frame(chromosome = c(paste0("Chr",1:10)),
                                 chr_length = c(301284077, 237032434, 232132502, 241403054, 217807808,
                                                169151979, 176396028, 175788001, 156637819, 150177341),
                                 fill = rep("slategray2",10),
                                 color = rep("black",10))
gene_info_default <- data.frame(chromosome = c("Chr3","Chr4","Chr8","Chr3","Chr6","Chr6","Chr4"),
                                start = c(1260877,233800183,158089073,210103369,3052152,88701559,82892492),
                                end = c(1262711,233811237,158099500,210107016,3054637,88711918,82893957),
                                label = c("wrky93","cts3","vln2","RINGLET 2","Rpl19","CRK23","myb28"),
                                gene_color = rep("black",7),
                                label_color = rep("red",7))
GMP(chr_length = chr_length_default,gene_info = gene_info_default)
}
